requirements.txt
requests
selenium
webdriver-manager
beautifulsoup4
ollama
python-dotenv 